feat(venue): extract useVenueSearchParameters from folder
