+++ src/features/venue/api/useVenueOffers.ts
@@ -4,7 +4,7 @@ import { useQuery } from 'react-query'
 import { VenueResponse } from 'api/gen'
 import { useIsUserUnderage } from 'features/profile/helpers/useIsUserUnderage'
 import { MAX_RADIUS } from 'features/search/helpers/reducer.helpers'
-import { useVenueSearchParameters } from 'features/venue/helpers/useVenueSearchParameters/useVenueSearchParameters'
+import { useVenueSearchParameters } from 'features/venue/helpers/useVenueSearchParameters'
 import { useSearchAnalyticsState } from 'libs/algolia/analytics/SearchAnalyticsWrapper'
 import { fetchOffers } from 'libs/algolia/fetchAlgolia/fetchOffers'
 import { useTransformOfferHits, filterOfferHit } from 'libs/algolia/fetchAlgolia/transformOfferHit'
