+++ src/features/venue/helpers/useNavigateToSearchWithVenueOffers.ts
@@ -2,7 +2,7 @@ import { VenueResponse } from 'api/gen'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
 import { TabParamList } from 'features/navigation/TabBar/types'
 import { SearchView } from 'features/search/types'
-import { useVenueSearchParameters } from 'features/venue/helpers/useVenueSearchParameters/useVenueSearchParameters'
+import { useVenueSearchParameters } from 'features/venue/helpers/useVenueSearchParameters'
 
 export const useNavigateToSearchWithVenueOffers = (venue: VenueResponse) => {
   const venueSearchParams: TabParamList['Search'] = useVenueSearchParameters(venue)
